+++ TopTagger/src/TTMHEPRequirements.cpp
@@ -43,13 +43,13 @@ void TTMHEPRequirements::run(TopTaggerResults& ttResults)
         bool passHEPRequirments = false;
 
         //Get the total candidate mass
-        double m123 = topCand.p().M();
+        float m123 = topCand.p().M();
 
         if(doTrijet_ && jets.size() == 3) //trijets
         {
-            double m12  = (jets[0]->p() + jets[1]->p()).M();
-            double m23  = (jets[1]->p() + jets[2]->p()).M();
-            double m13  = (jets[0]->p() + jets[2]->p()).M();
+            float m12  = (jets[0]->p() + jets[1]->p()).M();
+            float m23  = (jets[1]->p() + jets[2]->p()).M();
+            float m13  = (jets[0]->p() + jets[2]->p()).M();
 
             //Implement HEP mass ratio requirements here
             bool criterionA = 0.2 < atan(m13/m12) &&
@@ -74,11 +74,11 @@ void TTMHEPRequirements::run(TopTaggerResults& ttResults)
         }
         else if(doDijet_ && jets.size() == 2) //dijets
         {
-            double m23  = (jets[0]->getType() == AK8JET)?(jets[0]->getSoftDropMass() * jets[0]->getWMassCorr()):(jets[1]->getSoftDropMass() * jets[1]->getWMassCorr());
+            float m23  = (jets[0]->getType() == AK8JET)?(jets[0]->getSoftDropMass() * jets[0]->getWMassCorr()):(jets[1]->getSoftDropMass() * jets[1]->getWMassCorr());
             //small hack for legacy tagger
             if(jets[0]->getType() == AK4JET && jets[1]->getType() == AK4JET) m23 = jets[0]->p().M();
 
-            double m123 = topCand.p().M();
+            float m123 = topCand.p().M();
             if(jets[0]->getType() == AK8JET)
             {
                 TLorentzVector psudoVec;
